+++ Rendering/vtkCamera.h
@@ -60,7 +60,9 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector3Macro(Position,double);
   void GetPosition(float a[3]) {
     double tmp[3]; this->GetPosition(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; a[2] = tmp[2]; };     
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); 
+    a[2] = static_cast<float>(tmp[2]); };     
 
   // Description:
   // Set/Get the focal of the camera in world coordinates.
@@ -73,7 +75,9 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector3Macro(FocalPoint,double);
   void GetFocalPoint(float a[3]) {
     double tmp[3]; this->GetFocalPoint(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; a[2] = tmp[2]; }; 
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); 
+    a[2] = static_cast<float>(tmp[2]); }; 
   
   // Description:
   // Set/Get the view up direction for the camera.  The default
@@ -86,7 +90,9 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector3Macro(ViewUp,double);
   void GetViewUp(float a[3]) {
     double tmp[3]; this->GetViewUp(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; a[2] = tmp[2]; }; 
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); 
+    a[2] = static_cast<float>(tmp[2]); }; 
 
   // Description:
   // Recompute the ViewUp vector to force it to be perpendicular to
@@ -107,7 +113,9 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector3Macro(DirectionOfProjection,double);
   void GetDirectionOfProjection(float a[3]) {
     double tmp[3]; this->GetDirectionOfProjection(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; a[2] = tmp[2]; }; 
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); 
+    a[2] = static_cast<float>(tmp[2]); }; 
 
   // Description:
   // Move the position of the camera along the direction of projection. Moving
@@ -206,7 +214,8 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector2Macro(ClippingRange,double);
   void GetClippingRange(float a[2]) {
     double tmp[2]; this->GetClippingRange(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; }; 
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); }; 
 
   // Description:
   // Set the distance between clipping planes.  This method adjusts the 
@@ -249,7 +258,9 @@ class VTK_RENDERING_EXPORT vtkCamera : public vtkObject
   vtkGetVector3Macro(ViewPlaneNormal,double);
   void GetViewPlaneNormal(float a[3]) {
     double tmp[3]; this->GetViewPlaneNormal(tmp); 
-    a[0] = tmp[0]; a[1] = tmp[1]; a[2] = tmp[2]; }; 
+    a[0] = static_cast<float>(tmp[0]); 
+    a[1] = static_cast<float>(tmp[1]); 
+    a[2] = static_cast<float>(tmp[2]); }; 
 
 
   // Description:
